refactor doozer olm_bundle to new file
